requirements.txt
ploupy-sdk
python-dotenv